test(cache): split inspector integration suites
